// src/generated/prisma/models/Campaign.ts
export type CampaignWhereInput = {
  AND?: Prisma.CampaignWhereInput | Prisma.CampaignWhereInput[]
  OR?: Prisma.CampaignWhereInput[]
  NOT?: Prisma.CampaignWhereInput | Prisma.CampaignWhereInput[]
  id?: Prisma.StringFilter<"Campaign"> | string
  name?: Prisma.StringFilter<"Campaign"> | string
  slug?: Prisma.StringFilter<"Campaign"> | string
  description?: Prisma.StringNullableFilter<"Campaign"> | string | null
  status?: Prisma.EnumCampaignStatusFilter<"Campaign"> | $Enums.CampaignStatus
  startsAt?: Prisma.DateTimeNullableFilter<"Campaign"> | Date | string | null
  endsAt?: Prisma.DateTimeNullableFilter<"Campaign"> | Date | string | null
  bannerUrl?: Prisma.StringNullableFilter<"Campaign"> | string | null
  thumbnailUrl?: Prisma.StringNullableFilter<"Campaign"> | string | null
  isActive?: Prisma.BoolFilter<"Campaign"> | boolean
  metaTitle?: Prisma.StringNullableFilter<"Campaign"> | string | null
  metaDescription?: Prisma.StringNullableFilter<"Campaign"> | string | null
  metaImageUrl?: Prisma.StringNullableFilter<"Campaign"> | string | null
  isFeatured?: Prisma.BoolFilter<"Campaign"> | boolean
  createdAt?: Prisma.DateTimeFilter<"Campaign"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Campaign"> | Date | string
  archivedAt?: Prisma.DateTimeNullableFilter<"Campaign"> | Date | string | null
  banners?: Prisma.CampaignBannerListRelationFilter
  sections?: Prisma.CampaignSectionListRelationFilter
  promotions?: Prisma.PromotionCampaignListRelationFilter
}